TaxiServlet/src/main/java/com/example/taxi/controller/commands/user/UserNewCommand.java
package com.example.taxi.controller.commands.user;

import com.example.taxi.controller.commands.Command;
import com.example.taxi.model.entity.User;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

public class UserNewCommand implements Command {
    @Override
    public String execute(HttpServletRequest request, HttpServletResponse response) throws ServletException {
        request.setAttribute("user", new User());
        request.setAttribute("mode", 0);
        return "/WEB-INF/jsp/user.jsp";
    }
}